Fixes according to comments, custom clustering var, moved to preprocess
